+++ src/syrag/__init__.py
@@ -196,16 +196,13 @@ else:
     __all__.extend(["GoogleEmbedder", "GoogleLLM"])
 
 try:
-    from syrag.integrations.langchain import LangChainTextChunker
+    from syrag.integrations.langchain import LangChainRetrieverStrategy, LangChainTextChunker
 except ModuleNotFoundError:
     pass
 else:
-    globals().update(
-        {
-            "LangChainTextChunker": LangChainTextChunker,
-        }
-    )
-    __all__.extend(["LangChainTextChunker"])
+    globals()["LangChainRetrieverStrategy"] = LangChainRetrieverStrategy
+    globals()["LangChainTextChunker"] = LangChainTextChunker
+    __all__.extend(["LangChainRetrieverStrategy", "LangChainTextChunker"])
 
 try:
     from syrag.testing import (