+++ src/syrag/integrations/__init__.py
@@ -1,9 +1,10 @@
 __all__: list[str] = []
 
 try:
-    from syrag.integrations.langchain import LangChainTextChunker
+    from syrag.integrations.langchain import LangChainRetrieverStrategy, LangChainTextChunker
 except ModuleNotFoundError:
     pass
 else:
-    globals().update({"LangChainTextChunker": LangChainTextChunker})
-    __all__.extend(["LangChainTextChunker"])
+    globals()["LangChainRetrieverStrategy"] = LangChainRetrieverStrategy
+    globals()["LangChainTextChunker"] = LangChainTextChunker
+    __all__.extend(["LangChainRetrieverStrategy", "LangChainTextChunker"])